Refactor RevisionPage component and rename files to match updated naming conventions Relates #12
